backend/requirements.txt
fastapi[standard]
uvicorn
pydantic
motor
python-decouple

